# Use `slices.SortFunc` in OpenCensus bridge instead of `sort` package