Avoids the allocation of the custom sortable types.


```console
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/bridge/opencensus/internal/ocmetric
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                   │   old.txt   │               new.txt               │
                   │   sec/op    │   sec/op     vs base                │
ConvertExemplar-8    2.113µ ± 4%   1.947µ ± 4%   -7.86% (p=0.001 n=10)
ConvertQuantiles-8   1.639µ ± 3%   1.454µ ± 4%  -11.26% (p=0.000 n=10)
geomean              1.861µ        1.683µ        -9.57%

                   │   old.txt    │               new.txt               │
                   │     B/op     │     B/op      vs base               │
ConvertExemplar-8    2.086Ki ± 0%   2.062Ki ± 0%  -1.12% (p=0.000 n=10)
ConvertQuantiles-8     384.0 ± 0%     360.0 ± 0%  -6.25% (p=0.000 n=10)
geomean                905.7          872.0       -3.72%

                   │  old.txt   │              new.txt               │
                   │ allocs/op  │ allocs/op   vs base                │
ConvertExemplar-8    6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=10)
ConvertQuantiles-8   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
geomean              4.243        3.162       -25.46%
```