@MrAlias 2024-02-28
> > Avoids the allocation of the custom sortable types.
> 
> Nice to have: benchstat output in the description 😉

Benchmarks added.